--- src/firebase/firebaseConfig.ts
@@ -1,16 +1,0 @@
-import Constants from "expo-constants";
-import { initializeApp } from "firebase/app";
-
-const firebaseConfig = {
-  apiKey: Constants.expoConfig?.extra?.firebase.apiKey,
-  authDomain: Constants.expoConfig?.extra?.firebase.authDomain,
-  projectId: Constants.expoConfig?.extra?.firebase.projectId,
-  storageBucket: Constants.expoConfig?.extra?.firebase.storageBucket,
-  messagingSenderId: Constants.expoConfig?.extra?.firebase.messagingSenderId,
-  appId: Constants.expoConfig?.extra?.firebase.appId,
-  measurementId: Constants.expoConfig?.extra?.firebase.measurementId,
-};
-
-const app = initializeApp(firebaseConfig);
-
-export default app;
