+++ src/screens/auth/login.tsx
@@ -6,7 +6,7 @@ import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import Spinner from "react-native-loading-spinner-overlay";
 
 import { AuthParamList } from "@navigation/navigationTypes";
-import { useAuth } from "@firebaseApi";
+import { useAuth } from "firebase/api";
 import CustomField from "components/form/formField";
 import { ContainerScroll as Container } from "@styledComponents";
 import { colors } from "@theme";
