+++ src/screens/home/post/createPost.tsx
@@ -1,12 +1,10 @@
 import { useState } from "react";
 import { View } from "react-native";
-import { Timestamp } from "firebase/firestore";
 import { useForm } from "react-hook-form";
 import { useUnit } from "effector-react";
-import { Button } from "react-native-paper";
 import Spinner from "react-native-loading-spinner-overlay";
-import { usePost } from "@firebaseApi";
-import { CreatePostProps, UpdatePostProps } from "@navigation/navigationTypes";
+import { usePost } from "firebase/api";
+import { CreatePostProps } from "@navigation/navigationTypes";
 import { $officeStore } from "@context/officeStore";
 import CustomField from "components/form/formField";
 import { ContainerScroll } from "@styledComponents";
