+++ src/screens/office/partnership/createPartnership.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useStoreMap, useUnit } from "effector-react";
 
-import { usePartnership } from "@firebaseApi";
+import { usePartnership } from "firebase/api";
 import { $officeStore } from "@context/officeStore";
 import { CreatePartnershipProps } from "@navigation/navigationTypes";
 import { FormFieldValues, PartnershipFieldNames } from "@types";
